# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/obligations": {
      "get": {
        "operationId": "ObligationsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 10000,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response. All records are demo data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationRecordPage"
                }
              }
            }
          },
          "400": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Obligations"
        ]
      },
      "post": {
        "operationId": "ObligationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateObligationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response. All records are demo data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationRecord"
                }
              }
            }
          },
          "400": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Obligations"
        ]
      }
    },
    "/api/v1/obligations/{id}": {
      "patch": {
        "operationId": "ObligationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateObligationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response. All records are demo data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationRecord"
                }
              }
            }
          },
          "400": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Obligations"
        ]
      },
      "delete": {
        "operationId": "ObligationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Obligations"
        ]
      }
    },
...
  },
  "info": {
    "title": "RASID API",
    "description": "Arabic-first backend portfolio. Synthetic/manual/sanitized demo data only. All money values are decimal strings. No banking, payments, investment execution or financial advice. See docs/FRONTEND_HANDOFF.md for browser integration and response examples.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "rasid_refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rasid_refresh"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreateObligationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "إيجار تجريبي"
          },
          "amount": {
            "type": "string",
            "example": "1800.00"
          },
          "currency": {
            "type": "string",
            "enum": [
              "SAR",
              "USD",
              "EUR"
            ]
          },
          "dueDay": {
            "type": "number",
            "minimum": 1,
            "maximum": 28,
            "example": 1
          },
          "categoryId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "isActive": {
            "type": "object",
            "default": true
          }
        },
        "required": [
          "name",
          "amount",
          "currency",
          "dueDay"
        ]
      },
      "UpdateObligationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "إيجار تجريبي"
          },
          "amount": {
            "type": "string",
            "example": "1800.00"
          },
          "currency": {
            "type": "string",
            "enum": [
              "SAR",
              "USD",
              "EUR"
            ]
          },
          "dueDay": {
            "type": "number",
            "minimum": 1,
            "maximum": 28,
            "example": 1
          },
          "categoryId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "messageAr": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {}
              }
            },
            "required": [
              "code",
              "message",
              "messageAr"
            ]
          },
          "requestId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "error",
          "requestId",
          "timestamp"
        ]
      },
...
      "ObligationRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "amount": {
            "type": "string",
            "example": "42.50",
            "description": "Exact decimal string, never a JSON floating-point number."
          },
          "currency": {
            "type": "string",
            "enum": [
              "SAR",
              "USD",
              "EUR"
            ]
          },
          "dueDay": {
            "type": "integer"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "createdAt",
          "userId",
          "name",
          "amount",
          "currency",
          "dueDay",
          "categoryId",
          "isActive"
        ]
      },
...
      "ObligationRecordPage": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObligationRecord"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            },
            "required": [
              "page",
              "limit",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      }
    }
  }
}